Rename togl to togls.
I did this because since we started this project it seems someone else has
created another feature toggle library with the same name. I dug into the
other library and we are a bit further along in terms of code quality and
flexibility and functionality. Therefore, I decide to just rename it and
compete.