Rename variable in FeatureRegistry.create
The variable `registry` was previously used in the
`FeatureRegistry.create` class method. This is a bit confusing when you
also notice that the `FeatureRegistry#registry` instance method also
exists. Beyond that, the `FeatureRegistry` class also has an instance
variables, `@registry`. All these things being called `registry` was
confusing so I renamed the `registry` variable to `feature_registry`
in the `FeatureRegistry.create` method to make it clearer that it is a
full `FeatureRegistry` class instance, not just in internal
representation of a registry.