Eliminate a useless `tap()` call
I eliminated the `tap()` call in this because the `Feature#on` method
returns `self`, which is the instance of the `Feature` itself.
Therefore, it would be the same as what the `tap()` call is doing. I
suspect this was done maybe before the `on()` method returned `self`.
Either way, now it seems useless to use a `tap()` call there.