Make default boolean rule klass configurable
I did this in preparation for adding environment variable based override
boolean rules. This way, we can simply change the default boolean rule that is
used depending on the app it is used in. Some teams may want the
BooleanEnvOverride rules that are coming in the future, while others may still
just want the normal Boolean rule to be default.