Switch Rule definition to pass key to run()
I did this so that our rules have another bit of information they can
use to evaluate their boolean return value. This is especially useful
for the new BooleanEnvOverrideRule as it needs to lookup an environment
variable based on the feature key. I also went through the README.md and
cleaned up a lot and corrected references and examples to clearly abide
by the new Rule interface.