Move outputting feature toggles to wiki
Why you made the change:

I did this because it isn't a critical feature to get going with the
library and therefore is contributing to the README.md being
overwhelming. Hence, I am shifting it to the Wiki so when people want to
dig a bit further into what is offered they can then discover it.