Rework README.md for wiki extraction
Why you made the change:

I did this because we recently extracted portions of the README.md to
the Wiki and I hadn't had a chance to rework the README.me to make more
sense with the Wiki.

How the change addresses the need:

This isn't necessarily complete. However, it is an attempt to move a
step closer.