Got environment override working
Why you made the change:

I did this so that people could override their toggles using environment
variables. Note: This is implemented with kind of a hack, as the driver
duplicates the mechanism of generating a Rule#id. If you have any better
ideas on how to implement this. Please don't hesitate.

How the change addresses the need: