Extracted Rule#id digest out into helper
**Why you made the change**

I did this because we had already duplicated the logic in a number of
places to support both the Rule#id as well as the
ToggleRepositoryDrivers::EnvOverrideDriver.

**How the change addresses the need**

Extracting this into a helper allowed me to replace all those disparate
implementations and replace them all with simple calls to this helper.
It also helps make sure that the mechanism of defining the unique
identifiers is consistent across the code.