Add togls monodraw architecture diagram
Why you made the change:

I did this so that people interested in contributing or simply learning
about the togls architecture would find it easier to get up to speed.

How the change addresses the need:

This will help with that by quickly letting them see the relationships
between the various architectural entities.