Switch .ruby-version to 2.2.3
**Why you made the change**

I did this so all the developers would be on the latest and greatest and
be on the same page.

**How the change addresses the need**

This change causing rbenv/rvm/other ruby version managers to push the
developer to switch into ruby 2.2.3 as the ruby version managers detect
this file and then try to switch into that version of ruby. If that
version of ruby is not installed the version managers push the developer
to install that version of ruby.