Switch from img.shields.io badges to native badges
**Why you made the change**

I did this because img.shields.io has been having problems lately and
hasn't been properly producing the badges. Beyond that, the need for
img.shields.io seems to have subsided because most of the providers are
providing SVGs directly now.

**How the change addresses the need**

Switching to the native badge providers distributes the badge failures
amongst each of the respective services. Beyond that, it is one less
layer to go through.