Hard wrapped the CODE_OF_CONDUCT.md
**Why you made the change**

I did this so that it would be easier to read in a text editor and
easier to edit.

**How the change addresses the need**

Having the lines wrap at just under 80 characters makes it much easier
to read as your eyes don't have to travel across the entire horizontal
axis of the screen and then have to track where the next line is.