Upgrade to the latest codeclimate-test-reporter
**Why you made the change**

I did this so that it would work best with the codeclimate service and
so that we are as up to date as possible.

**How the change addresses the need**

I did not have a specific failure/bug case that I made this update for.
I simply updated it so that we would have the latest and greatest
codeclimate support.