Remove the BooleanEnvOverride rule
**Why you made the change**

I removed this rule because it's functionality has been replaced with
the Togls::ToggleRepositoryDrivers::EnvOverrideDriver in the new
architecture.

**How the change addresses the need**

Removing this has zero functional impact as it was previously replaced
with the functionality of the
Togls::ToggleRepositoryDrivers::EnvOverrideDriver implementation based
on the new architecture. Therefore, this is simply a cleanup of unused
code in a hopes to remove confusion.