Add an `off?` conditional method for features
I did this because I want a way to ask if a feature is `off?` not just `on?`.
This is a little sugar but it will be helpful in some situations. I see this
as helpful when you are changing behavior.

Some examples would be:

```ruby
do_some_stuff if Togls.feature(:new_feature).off? && some_other_conditional

if Togls.feature(:new_feature).off?
 do_old_stuff
end
```