Clean up feature specs and organization
**Why you made the change**

I did this so that the feature specs would be more consumable both in
documentation form but also in development as well.

**How the change addresses the need**

Moving the "defaults to false when feature is not defined" spec into the
"Togl inspect" grouping makes more sense as that is really what that
test is about.

Adding the alternate case for environment variable overrides just makes
sure we have complete coverage. Changing the phrasing of the environment
variable override it's and context's help provide a clearer message when
run in documentation mode as well as within the specs.