Cleanup some APIs
Why you made the change:

I did this to make the software more maintainable by simplifying its
internals a bit as well as enabling the creation of empty feature toggle
sets..

How the change addresses the need:

Specifically I changed the Togls.features mechanism to no longer raise
an error when calling Togls.features and not providing a block. Instead
it now creates an empty toggle registry in the scenario where there
isn't already a toggle registry and expands that toggle registry with
the block when given.