Make Togls.feature create empty toggle registry
Why you made the change:

I did so that there wouldn't be a way from the external API screw
yourself by having a nil feature toggle registry. More specifically,
prior to this if you were to set `Togls.features = nil` and then try and
evaluate a feature toggle it would throw an exception, :-). This
resolves issue #19.