Bump version to v2.1.0
Why you made the change:

I did this because we fixed issue #19. Granted this changes the behavior
of one of the public interfaces from throwing an exception to creating
an empty feature toggle set, I opted to treat this as a backwards
compatible feature because I knew that the exception wasn't being used
in our public interface.