Make env override use true/false
Why you made the change:

I did this so that the only valid values for an environment variable
override are true/false. This means that any other value than true or
false will now result in the feature toggle falling back to the in
memory value for that feature toggle. Issue #21.