Add starter DEVELOPMENT.md
Why you made the change:

I did this so that people that wanted to contribute to this project
would have somewhere to go to find information to aid them in
understanding and contributing to the code.

How the change addresses the need:

This provides them details on where they should start and what they
should focus on initially (being end user experience and docs), as well
as where they should focus secondarily (understanding the architecture
within the context of the end user experience).