Add Togls::TestToggleRegistry class
Why you made the change:

I added this class specifically so that we would have a toggle registry
that didn't use environment override drivers. The is especially
important in the automated test suite scenario as identified by
issue #24. It basically prevents environment variables from interfering
with the explicitly configured toggle states in your automated tests.

How the change addresses the need:

It does so by allowing you in your test setup to set the toggle registry
to an instance of the new Togls::TestToggleRegistry class.