Make Togls drivers thread-safe
Why you made the change:

I did this so that Togls could be safely used in a threaded environment. I prime
example of this would be a web application run via puma.

How the change addresses the need:

The only real modifiable state is stored within the drivers so we didn't need
to add any synchronization above the repository driver layers.