Add Marshal to in-memory driver set, get, all
Why you made the change:

I did this because there was a rare timing issue with multi-threading where
references to sub-hashes and sub-values of a hash that were of certain types
where being incorrectly shared.

How the change addresses the need:

Adding the Marshaling in the set, get, and all methods of the drivers resolves
this issue because it basically causes a deep copy of the objects. It also has
the added benefit of matching the behavior of say the Redis driver and most
others that would also require marshaling.