Remove registry create method
Why you made the change:

I did this because it was duplicative of the responsibility of building
a registry. I modified the initialize method of both the
FeatureToggleRegistry class and the TestToggleRegistry class to accept
on optional block so that they can be used in a similar fashion to how
create was previously used. This also will allow a registry to have
arguments during construction in addition to the block that created
didn't really allow for.