Add ability to store target_type as metadata
Why you made the change:

I did this so that that one, the meta data would actually be able to be stored
and retrieved. Two, I did this so that the target_type meta data would be stored
as well so that in the future we would be able to support some degree of type
safety between rules and features using the target_type contract.