Add optional abstract target types
Why you made the change:

I did this so that we could define abstract target types when defining a feature
and also specify them when defining custom rules. This in turn allows us to
check for target type miss-matches and fall back to null based toggle that falls
back to a default boolean false. It also facilitates any servers implemented as
drivers of the repositories to be able to implement their own type checking
server side. This relates to issue #65.