Add ability for Rules to have a target_type
Why you made the change:

I did this because I want to add support for in code and externally defined
rules to be able to specify a given target_type. That way for example if you
were defining a rule instance called alpha_testers from the Group rule you would
be able to type match appropriately. For example:

alpha_testers = Togls::Rules::Group.new([323, 11, 34, 123], target_type: :user_id)

This would allow us to do checking based on the instance which if it doesn't
have a target_type falls back to the class (a.k.a. rule type). If we were to do
the following:

alpha_testers = Togls::Rules::Group.new([323, 11, 34, 123])

Then it would simply be a rule with a target_type of ANY which doesn't really
bring any additional safety in terms of these abstract formalized target types.

Note: This change is just in preparation for this is support. It does not modify
the type checking to use this, nor does it support reconstitution of rules with
an optional target_type. Both of these things would have to happen in order to
support this concept fully.