Add rule reconstitution with target_type
Why you made the change:

I did this so that if any of the rule repository drivers returned a target_type
then it would be used to construct the associated rule when it is reconstituted.
If no target_type is present then it will simply construct the rule without one,
in turn having that rule default to the rule type's target_type.