Modify target type, type checking to use rule target type
Why you made the change:

I did this because the necessary dependencies to support this functionality have
been put in place in a backwards compatible fashion. This will enable users to
specify target types specific to a given rule instance optionally. If it isn't
specified on the rule instance the rule instance target type will fall back to
the target type of that rule instances rule type.