Update mismatch logic for rule and feature target type check
I did this so that we can verify that the rule and feature can be evaluated
properly. If there is a mismatch it will return false and later return a type
of null toggle. This requires us to change the boolean rule to have the
correct default target type of NONE instead of being NOT_SET.