Validate target against feature target_type contract
I did this so that when a feature is evaluated in code developers can be
informed before deployment when there is a feature target type contract
breach.

This compares the evaluation target against the expectations of the features
target type contract and raises appropriate exceptions in the contract breach
scenarios.