Handle invalid rule data in toggle reconstitute
Why you made the change:

I did this so that if for whatever reason a repository driver is miss-behaving
and returning invalid payloads that it would identify it and notify via
exception so that the toggle reconstitution process can handle the scenario
appropriately by constructing and returning a null toggle. It does this
explicitly so that in these scenarios the application won't crash, instead the
feature being evaluated will simply be off.