Add type_id as a required param to a rule
Why you made the change:

I did this because the Rule instance needs to understand its abstract type
identifier in addition to its concrete type which it already understood. I
updated the Rule class and the specs around the Rule class. However, I did not
update the rest of the lib yet, so there is still work todo.