Add support for Ruby 2.4.0
Why you made the change:

I did this so that people would be able to use it with Ruby 2.4.0
based applications. The only required change was to fix a conflict
between changes made to Marhsal in 2.4.0 and RSpec. Therefore, I changed
the specs to no longer inject doubles such that Marshal would be called
on them. I also, added Ruby 2.4.0 to the .travis.yml so that we can get
our CI test suite running against Ruby 2.4.0 as well.